Remove old media player
When migrating this website from replit, i ran into a couple problems, one of which being that GitHub had file size limits, so i wasn't able to import one of the tracks i use a background music. I forgot about this, and today I finally remembered, and removed the media player that played that file, as it was just sitting there, broken, and forgotten. Have a great day :)